classe bouteille

+++ TP2_Bieres_BIGOT/src/tp2_bieres_bigot/BouteilleBiere.java
@@ -10,7 +10,7 @@ package tp2_bieres_bigot;
  */
 public class BouteilleBiere {
     String nom;
-    float degreAlcool;
+    double degreAlcool;
     String brasserie;
     boolean ouverte;  
     public void lireEtiquette() {
